# docs/swagger.yaml
basePath: /api
definitions:
  apperror.AppError:
    properties:
      message:
        type: string
    type: object
...
  brand.Brand:
    properties:
      banner:
        type: string
      country:
        $ref: '#/definitions/country.Country'
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      logo:
        type: string
      marketSection:
        $ref: '#/definitions/marketsection.MarketSection'
      marketSections:
        items:
          $ref: '#/definitions/marketsection.MarketSection'
        type: array
      name:
        type: string
      phoneNumber:
        type: string
      states:
        items:
          $ref: '#/definitions/state.State'
        type: array
      updatedAt:
        type: string
    type: object
  brand.BrandSummary:
    properties:
      id:
        type: integer
      logo:
        type: string
      name:
        type: string
    type: object
  country.Country:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  handler.BrandRequest:
    properties:
      banner:
        type: string
      country:
        type: integer
      email:
        type: string
      logo:
        type: string
      marketSection:
        type: integer
      marketSubSectionIds:
        items:
          type: integer
        type: array
      name:
        type: string
      phoneNumber:
        type: string
      stateIds:
        items:
          type: integer
        type: array
    required:
    - banner
    - country
    - email
    - logo
    - marketSection
    - marketSubSectionIds
    - name
    - phoneNumber
    - stateIds
    type: object
  handler.BrandResponse:
    properties:
      brand:
        $ref: '#/definitions/brand.Brand'
    type: object
  handler.BrandsSummaryResponse:
    properties:
      brands:
        items:
          $ref: '#/definitions/brand.BrandSummary'
        type: array
    type: object
  handler.BusinessProfileRequest:
    properties:
      businessIndustryId:
        type: integer
      businessName:
        maxLength: 30
        minLength: 3
        type: string
      countryId:
        type: integer
      email:
        type: string
      phoneNumber:
        type: string
      regionId:
        type: integer
      stateId:
        type: integer
    required:
    - businessIndustryId
    - businessName
    - countryId
    - email
    - phoneNumber
    - regionId
    - stateId
    type: object
...
  handler.FileResponse:
    properties:
      file:
        $ref: '#/definitions/upload.File'
    type: object
...
  handler.ProfileRequest:
    properties:
      age:
        type: integer
      countryId:
        type: integer
      firstName:
        maxLength: 30
        minLength: 3
        type: string
      lastName:
        maxLength: 30
        minLength: 3
        type: string
      phoneNumber:
        type: string
      regionId:
        type: integer
      stateId:
        type: integer
    type: object
...
  marketsection.MarketSection:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  region.Region:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  state.State:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  upload.File:
    properties:
      contentType:
        type: string
      name:
        type: string
      size:
        type: integer
    type: object
  user.BusinessIndustry:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  user.BusinessProfile:
    properties:
      businessIndustry:
        $ref: '#/definitions/user.BusinessIndustry'
      businessName:
        type: string
      country:
        $ref: '#/definitions/country.Country'
      email:
        type: string
      phoneNumber:
        type: string
      region:
        $ref: '#/definitions/region.Region'
      state:
        $ref: '#/definitions/state.State'
    type: object
  user.BusinessProfileResponse:
    properties:
      businessProfile:
        $ref: '#/definitions/user.BusinessProfile'
    type: object
  user.User:
    properties:
      age:
        type: integer
      avatar:
        type: string
      country:
        $ref: '#/definitions/country.Country'
      email:
        type: string
      firstName:
        type: string
      hasBusinessProfile:
        type: boolean
      id:
        type: integer
      isAdmin:
        type: boolean
      isPasswordSet:
        type: boolean
      isVerified:
        type: boolean
      lastName:
        type: string
      phoneNumber:
        type: string
      region:
        $ref: '#/definitions/region.Region'
      state:
        $ref: '#/definitions/state.State'
      username:
        type: string
    type: object
  user.UserResponse:
    properties:
      user:
        $ref: '#/definitions/user.User'
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for Kushfinds application
  title: Kushfinds API
  version: "1.0"
paths:
...
  /me/brands:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.BrandsSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - ApiKeyAuth: []
      tags:
      - market
    post:
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.BrandRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.BrandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - ApiKeyAuth: []
      tags:
      - market
  /me/brands/{id}:
    delete:
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - ApiKeyAuth: []
      tags:
      - market
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.BrandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - ApiKeyAuth: []
      tags:
      - market
    patch:
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.BrandRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.BrandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - ApiKeyAuth: []
      tags:
      - market
...
  /static/{filename}:
    get:
      parameters:
      - description: file name
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      tags:
      - upload
  /upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: form data
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.FileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - ApiKeyAuth: []
      tags:
      - upload
  /users/business:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.BusinessProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - ApiKeyAuth: []
      tags:
      - users
    patch:
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.BusinessProfileRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.BusinessProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - ApiKeyAuth: []
      tags:
      - users
  /users/me:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - ApiKeyAuth: []
      tags:
      - users
  /users/profile:
    patch:
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ProfileRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - ApiKeyAuth: []
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"